Make use of enums in thermostat properties Break out tests to a more managable size Add more tests Remove legacy imports and methods no longer used by Home Assistant
