+++ tests/sensors/__init__.py
@@ -1,0 +1,4 @@
+"""Tests for pydeCONZ sensors.
+
+pytest --cov-report term-missing --cov=pydeconz tests/sensors
+"""
