+++ tests/test_sensors.py
@@ -10,23 +10,9 @@ import pytest
 from pydeconz.interfaces.sensors import (
     SwitchDeviceMode,
     SwitchMode,
-    ThermostatFanMode,
-    ThermostatMode,
-    ThermostatPreset,
-    ThermostatSwingMode,
-    ThermostatTemperatureMeasurement,
     SwitchWindowCoveringType,
 )
 from pydeconz.models.sensor.switch import SWITCH_DEVICE_MODE_DUAL_ROCKER
-from pydeconz.models.sensor.thermostat import (
-    THERMOSTAT_FAN_MODE_AUTO,
-    THERMOSTAT_MODE_AUTO,
-    THERMOSTAT_MODE_OFF,
-    THERMOSTAT_PRESET_AUTO,
-    THERMOSTAT_SWING_MODE_HALF_OPEN,
-    THERMOSTAT_TEMPERATURE_MEASUREMENT_MODE_FLOOR_SENSOR,
-    Thermostat,
-)
 
 
 @pytest.fixture
@@ -1477,346 +1463,6 @@ async def test_temperature_sensor(deconz_sensor):
     assert sensor.software_version == "20161129"
     assert sensor.type == "ZHATemperature"
     assert sensor.unique_id == "00:15:8d:00:02:45:dc:53-01-0402"
-
-
-async def test_configure_thermostat(
-    mock_aioresponse, deconz_session, deconz_called_with
-):
-    """Verify that configuring presence sensor works."""
-    thermostat = deconz_session.sensors.thermostat
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await thermostat.set_config(
-        id="0",
-        cooling_setpoint=1000,
-        enable_schedule=True,
-        external_sensor_temperature=24,
-        external_window_open=True,
-        fan_mode=ThermostatFanMode.AUTO,
-        flip_display=False,
-        heating_setpoint=500,
-        locked=True,
-        mode=ThermostatMode.AUTO,
-        mounting_mode=False,
-        on=True,
-        preset=ThermostatPreset.AUTO,
-        schedule=[],
-        set_valve=True,
-        swing_mode=ThermostatSwingMode.HALFOPEN,
-        temperature_measurement=ThermostatTemperatureMeasurement.FLOORSENSOR,
-        window_open_detection=True,
-    )
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={
-            "coolsetpoint": 1000,
-            "schedule_on": True,
-            "externalsensortemp": 24,
-            "externalwindowopen": True,
-            "fanmode": "auto",
-            "displayflipped": False,
-            "heatsetpoint": 500,
-            "locked": True,
-            "mode": "auto",
-            "mountingmode": False,
-            "on": True,
-            "preset": "auto",
-            "schedule": [],
-            "setvalve": True,
-            "swingmode": "half open",
-            "temperaturemeasurement": "floor sensor",
-            "windowopen_set": True,
-        },
-    )
-
-
-async def test_danfoss_thermostat(deconz_sensor):
-    """Verify that Danfoss thermostat works.
-
-    Danfoss thermostat is the simplest kind with only control over temperaturdeconz_sensore.
-    """
-    sensor = await deconz_sensor(
-        {
-            "config": {
-                "battery": 59,
-                "displayflipped": None,
-                "heatsetpoint": 2100,
-                "locked": None,
-                "mountingmode": None,
-                "offset": 0,
-                "on": True,
-                "reachable": True,
-            },
-            "ep": 1,
-            "etag": "6130553ac247174809bae47144ee23f8",
-            "lastseen": "2020-11-29T19:31Z",
-            "manufacturername": "Danfoss",
-            "modelid": "eTRV0100",
-            "name": "Thermostat_stue_sofa",
-            "state": {
-                "errorcode": None,
-                "lastupdated": "2020-11-29T19:28:40.665",
-                "mountingmodeactive": False,
-                "on": True,
-                "temperature": 2102,
-                "valve": 24,
-                "windowopen": "Closed",
-            },
-            "swversion": "01.02.0008 01.02",
-            "type": "ZHAThermostat",
-            "uniqueid": "14:b4:57:ff:fe:d5:4e:77-01-0201",
-        },
-    )
-
-    assert sensor.ZHATYPE == ("ZHAThermostat", "CLIPThermostat")
-
-    assert sensor.cooling_setpoint is None
-    assert sensor.display_flipped is None
-    assert sensor.error_code is None
-    assert sensor.external_sensor_temperature is None
-    assert sensor.external_window_open is None
-    assert sensor.fan_mode is None
-    assert sensor.floor_temperature is None
-    assert sensor.heating is None
-    assert sensor.heating_setpoint == 21.00
-    assert sensor.locked is None
-    assert sensor.mode is None
-    assert sensor.mounting_mode is None
-    assert sensor.mounting_mode_active is False
-    assert sensor.offset == 0
-    assert sensor.preset is None
-    assert sensor.state_on
-    assert sensor.swing_mode is None
-    assert sensor.temperature == 2102
-    assert sensor.scaled_temperature == 21.0
-    assert sensor.temperature_measurement is None
-    assert sensor.valve == 24
-    assert sensor.window_open_detection is None
-
-    # DeconzSensor
-    assert sensor.battery == 59
-    assert sensor.ep == 1
-    assert sensor.low_battery is None
-    assert sensor.on is True
-    assert sensor.reachable is True
-    assert sensor.tampered is None
-    assert sensor.secondary_temperature is None
-
-    # DeconzDevice
-    assert sensor.deconz_id == "/sensors/0"
-    assert sensor.etag == "6130553ac247174809bae47144ee23f8"
-    assert sensor.manufacturer == "Danfoss"
-    assert sensor.model_id == "eTRV0100"
-    assert sensor.name == "Thermostat_stue_sofa"
-    assert sensor.software_version == "01.02.0008 01.02"
-    assert sensor.type == "ZHAThermostat"
-    assert sensor.unique_id == "14:b4:57:ff:fe:d5:4e:77-01-0201"
-
-
-async def test_eurotronic_thermostat(deconz_sensor):
-    """Verify that thermostat sensor works."""
-    sensor = await deconz_sensor(
-        {
-            "config": {
-                "battery": 100,
-                "displayflipped": True,
-                "heatsetpoint": 2100,
-                "locked": False,
-                "mode": "auto",
-                "offset": 0,
-                "on": True,
-                "reachable": True,
-            },
-            "ep": 1,
-            "etag": "25aac331bc3c4b465cfb2197f6243ea4",
-            "manufacturername": "Eurotronic",
-            "modelid": "SPZB0001",
-            "name": "Living Room Radiator",
-            "state": {
-                "lastupdated": "2019-02-10T22:41:32",
-                "on": False,
-                "temperature": 2149,
-                "valve": 0,
-            },
-            "swversion": "15181120",
-            "type": "ZHAThermostat",
-            "uniqueid": "00:15:8d:00:01:92:d2:51-01-0201",
-        },
-    )
-
-    assert sensor.ZHATYPE == ("ZHAThermostat", "CLIPThermostat")
-
-    assert sensor.cooling_setpoint is None
-    assert sensor.error_code is None
-    assert sensor.fan_mode is None
-    assert sensor.floor_temperature is None
-    assert sensor.heating is None
-    assert sensor.heating_setpoint == 21.00
-    assert sensor.locked is False
-    assert sensor.mode == "auto"
-    assert sensor.mounting_mode is None
-    assert sensor.mounting_mode_active is None
-    assert sensor.offset == 0
-    assert sensor.preset is None
-    assert not sensor.state_on
-    assert sensor.swing_mode is None
-    assert sensor.temperature == 2149
-    assert sensor.scaled_temperature == 21.5
-    assert sensor.temperature_measurement is None
-    assert sensor.valve == 0
-    assert sensor.window_open_detection is None
-
-    # DeconzSensor
-    assert sensor.battery == 100
-    assert sensor.ep == 1
-    assert sensor.low_battery is None
-    assert sensor.on is True
-    assert sensor.reachable is True
-    assert sensor.tampered is None
-    assert sensor.secondary_temperature is None
-
-    # DeconzDevice
-    assert sensor.deconz_id == "/sensors/0"
-    assert sensor.etag == "25aac331bc3c4b465cfb2197f6243ea4"
-    assert sensor.manufacturer == "Eurotronic"
-    assert sensor.model_id == "SPZB0001"
-    assert sensor.name == "Living Room Radiator"
-    assert sensor.software_version == "15181120"
-    assert sensor.type == "ZHAThermostat"
-    assert sensor.unique_id == "00:15:8d:00:01:92:d2:51-01-0201"
-
-
-async def test_tuya_thermostat(mock_aioresponse, deconz_sensor, deconz_called_with):
-    """Verify that Tuya thermostat works."""
-    sensor = await deconz_sensor(
-        {
-            "config": {
-                "battery": 100,
-                "heatsetpoint": 1550,
-                "locked": None,
-                "offset": 0,
-                "on": True,
-                "preset": "auto",
-                "reachable": True,
-                "schedule": {},
-                "schedule_on": None,
-                "setvalve": True,
-                "windowopen_set": True,
-            },
-            "ep": 1,
-            "etag": "36850fc8521f7c23606c9304b2e1f7bb",
-            "lastseen": "2020-11-11T21:23Z",
-            "manufacturername": "_TYST11_kfvq6avy",
-            "modelid": "fvq6avy",
-            "name": "fvq6avy",
-            "state": {"lastupdated": "none", "on": None, "temperature": 2290},
-            "swversion": "20180727",
-            "type": "ZHAThermostat",
-            "uniqueid": "bc:33:ac:ff:fe:47:a1:95-01-0201",
-        },
-    )
-
-    assert sensor.ZHATYPE == Thermostat.ZHATYPE
-
-    assert sensor.heating_setpoint == 15.50
-    assert sensor.locked is None
-    assert sensor.mode is None
-    assert sensor.offset == 0
-    assert sensor.schedule_enabled is None
-    assert not sensor.state_on
-    assert sensor.temperature == 2290
-    assert sensor.scaled_temperature == 22.9
-    assert sensor.valve is None
-
-    # DeconzSensor
-    assert sensor.battery == 100
-    assert sensor.ep == 1
-    assert sensor.low_battery is None
-    assert sensor.on is True
-    assert sensor.reachable is True
-    assert sensor.tampered is None
-    assert sensor.secondary_temperature is None
-
-    # DeconzDevice
-    assert sensor.deconz_id == "/sensors/0"
-    assert sensor.etag == "36850fc8521f7c23606c9304b2e1f7bb"
-    assert sensor.manufacturer == "_TYST11_kfvq6avy"
-    assert sensor.model_id == "fvq6avy"
-    assert sensor.name == "fvq6avy"
-    assert sensor.software_version == "20180727"
-    assert sensor.type == "ZHAThermostat"
-    assert sensor.unique_id == "bc:33:ac:ff:fe:47:a1:95-01-0201"
-
-    # Verify temperature conversion to increase coverage
-    sensor.update(
-        {
-            "config": {
-                "coolsetpoint": 1000,
-                "externalsensortemp": 2000,
-                "heatsetpoint": None,
-            },
-            "state": {
-                "floortemperature": 4000,
-            },
-        }
-    )
-    assert sensor.cooling_setpoint == 10
-    assert sensor.external_sensor_temperature == 20
-    assert sensor.heating_setpoint is None
-    assert sensor.floor_temperature == 40
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await sensor.set_config(
-        cooling_setpoint=1000,
-        enable_schedule=True,
-        external_sensor_temperature=24,
-        external_window_open=True,
-        fan_mode=THERMOSTAT_FAN_MODE_AUTO,
-        flip_display=False,
-        heating_setpoint=500,
-        locked=True,
-        mode=THERMOSTAT_MODE_AUTO,
-        mounting_mode=False,
-        on=True,
-        preset=THERMOSTAT_PRESET_AUTO,
-        schedule=[],
-        set_valve=True,
-        swing_mode=THERMOSTAT_SWING_MODE_HALF_OPEN,
-        temperature_measurement=THERMOSTAT_TEMPERATURE_MEASUREMENT_MODE_FLOOR_SENSOR,
-        window_open_detection=True,
-    )
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={
-            "coolsetpoint": 1000,
-            "schedule_on": True,
-            "externalsensortemp": 24,
-            "externalwindowopen": True,
-            "fanmode": "auto",
-            "displayflipped": False,
-            "heatsetpoint": 500,
-            "locked": True,
-            "mode": "auto",
-            "mountingmode": False,
-            "on": True,
-            "preset": "auto",
-            "schedule": [],
-            "setvalve": True,
-            "swingmode": "half open",
-            "temperaturemeasurement": "floor sensor",
-            "windowopen_set": True,
-        },
-    )
-
-    mock_aioresponse.put("http://host:80/api/apikey/sensors/0/config")
-    await sensor.set_config(mode=THERMOSTAT_MODE_OFF)
-    assert deconz_called_with(
-        "put",
-        path="/sensors/0/config",
-        json={"mode": "off"},
-    )
 
 
 async def test_time_sensor(deconz_sensor):
